refactor(runtime): tighten public surface and colocate module tests
Move the plugin pipeline under session/plugins, relocate execution and
root API contract tests beside their modules, and trim execution host
exports to the aggregate durable contracts.
