+++ packages/runtime/src/execution/run/stored-agent-run.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
-import { createInMemoryExecutionHost } from "./execution/memory";
-import { StoredAgentRun } from "./execution/run/stored-agent-run";
-import type { AgentEvent } from "./session/protocol/events";
+import type { AgentEvent } from "../../session/protocol/events";
+import { createInMemoryExecutionHost } from "../memory";
+import { StoredAgentRun } from "./stored-agent-run";
 
 async function collectRunEvents(run: StoredAgentRun): Promise<AgentEvent[]> {
   const events: AgentEvent[] = [];
